+++ src/main/java/com/kh/vira_dev/ecommerceapi/dto/response/OrderCancelResponse.java
@@ -14,7 +14,7 @@ import java.time.Instant;
 @Builder
 public class OrderCancelResponse {
 
-    private Short orderId;
+    private Long orderId;
 
     private String orderCode;
