+++ src/main/java/com/kh/vira_dev/ecommerceapi/dto/response/PaymentSummaryResponse.java
@@ -16,7 +16,7 @@ import java.time.Instant;
 @Builder
 public class PaymentSummaryResponse {
 
-    private Integer paymentId;
+    private Long paymentId;
 
     private PaymentStatus paymentStatus;
