+++ src/main/java/com/kh/vira_dev/ecommerceapi/repository/CartItemRepository.java
@@ -3,5 +3,5 @@ package com.kh.vira_dev.ecommerceapi.repository;
 import com.kh.vira_dev.ecommerceapi.entity.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Short> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 }
