+++ src/main/java/com/kh/vira_dev/ecommerceapi/repository/CategoryRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category , Short> {
+public interface CategoryRepository extends JpaRepository<Category , Long> {
 
     boolean existsByName(String name);
