+++ src/main/java/com/kh/vira_dev/ecommerceapi/repository/ProductRepository.java
@@ -4,5 +4,5 @@ import com.kh.vira_dev.ecommerceapi.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ProductRepository extends JpaRepository<Product , Integer>, JpaSpecificationExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product , Long>, JpaSpecificationExecutor<Product> {
 }
