+++ src/main/java/com/kh/vira_dev/ecommerceapi/repository/ShippingAddressRepository.java
@@ -3,5 +3,5 @@ package com.kh.vira_dev.ecommerceapi.repository;
 import com.kh.vira_dev.ecommerceapi.entity.ShippingAddress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShippingAddressRepository extends JpaRepository<ShippingAddress , Short> {
+public interface ShippingAddressRepository extends JpaRepository<ShippingAddress , Long> {
 }
